Remove add_file fallback that can never run
The original intent of this block of code is to handle arguments such as
"politics-o", which would resolve to the "politics" fortune under the
offensive directories. This is a documented feature.

However, form_file_list() almost never calls add_file() with a file
argument without slashes -- if it's not definitely an absolute/relative
path, we always prepend $PWD, FORTDIR, or LOCFORTDIR, all of which are
absolute paths. This means that this condition wouldn't run in almost
all cases.

This feature will be re-added later in a way that is more functional.